Bug fix in Benchmark and consistent file naming
Consistent naming of file and folder names.
Benchmark removed from demo project as there is now a dedicated benchmark project.
Bug fixed in benchmark: SymSpell() instantiation now right before LoadDirectory(). maxLength was  shared between different dictionaries, this led to wrong result numbers.
SymSpell.defaultVerbose removed to force people to use and think about verbose, and to tie this parameter to Lookup where it is needed (verbose is only needed in Lookup, not in precalculation).
Add check to Lookup() whether verbose is in range 0..2
In lookup() the order of verbose and maxEditDistance parameters switched,. Verbose made obligatory and maxEditDistance optionally.